function moved to pkg package
